+++ src/player/index.js
@@ -1,2 +1,3 @@
 export { default as createPlayer } from './createPlayer';
 export { default as drawPlayer } from './drawPlayer';
+export { default as updatePlayer } from './updatePlayer';
