+++ src/utils/index.js
@@ -3,3 +3,6 @@ export { default as calculateRotation } from './calculateRotation';
 export { default as degreesToRadians } from './degreesToRadians';
 export { default as getRandomInt } from './getRandomInt';
 export { default as transformMapCoordinates } from './transformMapCoordinates';
+export { default as createCanvas } from './createCanvas';
+export { default as createDevbox } from './createDevbox';
+export { default as loadAssets } from './loadAssets';
