+++ src/utils/loadAssets.js
